chore: inline SchemaAndDescription
