+++ packages/elements-core/src/__fixtures__/articles/kitchen-sink.md
@@ -152,7 +152,85 @@ be the json schema object to be rendered.
 
 ```json
 {
-  "$ref": "https://stoplight.io/api/nodes.raw?srn=gh/stoplightio/studio-demo/reference/todos/models/todo-full.json"
+  "title": "Todo Full",
+  "allOf": [
+    {
+      "title": "Todo Partial",
+      "type": "object",
+      "properties": {
+        "name": {
+          "type": "string"
+        },
+        "completed": {
+          "type": [
+            "boolean",
+            "null"
+          ]
+        }
+      },
+      "required": [
+        "name",
+        "completed"
+      ],
+      "x-tags": [
+        "Todos"
+      ]
+    },
+    {
+      "type": "object",
+      "properties": {
+        "id": {
+          "type": "integer",
+          "minimum": 0,
+          "maximum": 1000000
+        },
+        "completed_at": {
+          "type": [
+            "string",
+            "null"
+          ],
+          "format": "date-time"
+        },
+        "created_at": {
+          "type": "string",
+          "format": "date-time"
+        },
+        "updated_at": {
+          "type": "string",
+          "format": "date-time"
+        },
+        "user": {
+          "title": "User",
+          "type": "object",
+          "properties": {
+            "name": {
+              "type": "string",
+              "description": "The user's full name."
+            },
+            "age": {
+              "type": "number",
+              "minimum": 0,
+              "maximum": 150
+            }
+          },
+          "required": [
+            "name",
+            "age"
+          ],
+          "x-tags": [
+            "Todos"
+          ]
+        }
+      },
+      "required": [
+        "id",
+        "user"
+      ]
+    }
+  ],
+  "x-tags": [
+    "Todos"
+  ]
 }
 ```
