+++ packages/elements-core/src/components/MarkdownViewer/index.tsx
@@ -12,3 +12,4 @@ export const MarkdownViewer = (props: IMarkdownViewerProps) => {
   return <MarkdownViewerComponent {...props} components={components} />;
 };
 MarkdownViewer.displayName = 'MarkdownViewer';
+export { SchemaAndDescription } from '@stoplight/elements-core/components/MarkdownViewer/CustomComponents/CodeComponent';
